+++ examples/example-status/src/test/java/de/otto/edison/example/ExampleStatusSmokeTest.java
@@ -3,7 +3,6 @@ package de.otto.edison.example;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.context.embedded.LocalServerPort;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.http.MediaType;
@@ -19,8 +18,6 @@ public class ExampleStatusSmokeTest {
 
     @Autowired
     private TestRestTemplate restTemplate;
-    @LocalServerPort
-    private int port;
 
     @Test
     public void shouldRenderMainPage() {
@@ -44,4 +41,9 @@ public class ExampleStatusSmokeTest {
         assertThat(response.getStatusCodeValue()).isIn(200, 503);
     }
 
+    @Test
+    public void shouldUseLdapAuthenticationForNonWhitelistedEndpoints() {
+        final ResponseEntity<String> response = this.restTemplate.getForEntity("/internal/jobs", String.class);
+        assertThat(response.getStatusCodeValue()).isEqualTo(401);
+    }
 }